Implement DAG
Implement a Direct Acyclic Graph (DAG) class that will serve as basis
for the report builder.

The underlying representation is a double linked tree (constructed by
transversing the `Node` objects of the graph). Nodes store both their
inputs and outputs as sets.

The nodes are identified and indexed by their value, which means that
there can only be one node with a given value in a graph. Also the
value should be *hashable* (e.g. no mutable list). This is OK for most
purposes (our nodes will be tuples of functions and parameters).

The graph is constructed by repeatedly calling `add_node(value,
inputs, outputs)` where the latter are sets containing the node
objects  or values that are inside the graph. This function wires up
the nodes, and checks for cycles.

A few common algorithms for iterating the nodes are implemented, the
most important being topological sort (which guarantees that
dependencies will be satisfied when executing nodes sequentially).

A dependency resolver generator is also implemented: Completed nodes
are sent to the generator and it returns the new nodes that have their
dependencies satisfied. This could serve to implement dependencies in
parallel.

Tests are implemented showing basic usage.