Check inputs and outputs before searching cycles
This optimizes for the common case where we add a node at the end of
the graph, and consequently cannot create cycles. While adding a node
is a (expensive) O(1) operation in the number of existing nodes,
checking for cycles is O(N), so it should be avoided when possible.