Add ExecModes
Add a generalization over the concept of mapping function names to
namespace keys. Specifically, the result of a function can now add an
element to a list if the `CallSpec` is marked with the
`APPEND_UNORDERED` mode (the name reminds that the order of execution
is not guaranteed as long as the requirements are met (and therefore
we can execute in parallel).

This allows for a useful generalization of what is written in the doc:
One could for example define *tests* such as  `test_outliers` and
`test_large_exponents` which are called optionally and append to
a single key `"tests"`. Then the plotting functions know how to
highlight the outliers that fail the tests (and do nothing if no tests
are provided).

The tests for execute_parallel fail for an unrelated reason which is
fixed in the next commit.