Don't use loop.stop()
Contrary to what the documentation implies `loop.stop()`, does not
wait for all the tasks to be completed before closing the loop but
rather seems to do it at the first available opportunity.

Instead of using it, we write everything as a single task and await
the futures instead of using callbacks. This is much more intuitive
and easier to follow.