Add utils module
This contains a `functools.partial` variant that compares equal with
partials for the same specification. It can also be used as a key for
dictionaries, with the key being computed by pickling the object
(which is the easiest and most general). This limits the values of
args and kwargs to pickable objects but that's the same limitation we
have for executing in parallel and should be enough for things that
can be passed inside the templates or configuration files, which is
the primary use case.