Implement template parsing
Create a custom environment for `jinja2` templates where the unknown
names in the template gets registered and the processed as targets.
This environment is a drop in replacement for the `jinja2` which it
extends, and can be used exactly in the same way. However when the
context manager `fetch_mode` is called, the values will get
substituted by dummy placeholders (on which all operations will fail)
and be registered and accessible as a list of targets. Then with
`subs_mode(results)`, the targets will get substituted by their
computed values. The interface is a little clumsy and a higher lever
wrapper should be produced, once the rest of the pieces are clear.

See test for details on behaviour.