Implement NS resolver
We change the paradigm, from having one namespace in total, to one
namespace per node.

With this several limitations of the current approach will be overcame
trivially (at the expense of more work).

The namespace where each node is to write is encoded in a namespace
specification (nsspec) so as to stack dictionaries, lists or elements
of lists within a given root dictionary (somehow initiated by the
config). In such a way, every complex element of the config can span
a namespace if required. The rules are:

 - Start with the keys of the root dictionary.
 - Specify the next element, which can be:
    * A dictionary: The keys of the inner dict are available as symbols, and if no key
      found, those of the outer).
    * A list: If the list elements are dictionaries, a namespace is
      spanned for each.
    * An element of a list: Does what you expect.

This commit breaks resourcebuilder, which will be heeavily modified.