Implement namespace functionality
This is the n-th attempt, but possibly the (roughly) good one. We
implement methods to resolve a "namespace spec" within a given
dictionary (namespace).

The spec consists on a tuple of "stack labels" which can be either
the key of a dictionary within a dictionary, or a tuple with a key and
an index, in the case the tuple is a list of dictionaries (in which
case the spec resolves to the element of that list).

There is also the concept of fuzzy specs, which are a list of labels,
which can correspond to either a dictionary or a list of dictionaries.
In the later case, they get resolved to every element of the list.

A generator based function is implemented in such a way that we can
construct the namespace as we process the fuzzy spec.

At this point, some TODOs are measuring and optimizing performance and
improving the interface based on the actual usage.