Make as_namespace resolution live in the namespace
This should work in all the case (though I am not sure). Previously
this depended from the parsed object, which is not what we want, as
only the first namespace of a map gets it's as_namespace value, and
the others simply miss it (or stack in an incorrect way).