Add simple check framework
The final form of this is to be decided, as it is a little too
complicated now. A check function is executed once all the
requirements for  the provider are satisfied, and takes the callspec,
and the current state of the namespace and the graph.

We use this to implement the require_one check which ensures that at
least one of the arguments is not null.