Create internal namespace for function defaults
Currently the function defaults pollute the namespace in which the
provider is required. This is not good because different providers
might have conflicting defaults, and also because resourcebuilder
relies on keys not been found, to search them in a different place.

Therefore we create a new namespace for each provider where we dump
the defaults. The changes required are:

 - It should not be counted for calculating the put_index, which
   should always ignore this defaults namespace and be >0.

 - The corresponding namespec should not propagate to several
   providers when they are chained. Thus we have to introduce the
  'back' parameter, which will roll the namespace back by one (the
   former default namespace) for the next provider.