Implement final_action protocol for figures
Instead of doing a rather ugly transformation to the input functions,
changing both their inputs and their semantics, we leave them
untouched, except for adding an attribute `final_action`. This will be
executed at the end of `set_result`, with the following parameters:

 result, spec, environment

This is to be revised (and should most probably contain the namespace
as well), but we leave it alone for now. In particular we want a way
of deriving a nicer filename than ($i).pdf.