Generate nice file names from figures
We change again the signature of `final_actions` (surely not for the
last time).

Instead of just using sequence numbers as filename, we now derive
paths that should be both nice and unique.

We still need a counter for figure generators (as they live in the
same namespec) but those could still be made nicer by returning a name
handle along with the figure (like in smpdf).