Rewrite process_requirement
Fix a problem when a resource is requested from two different
namespecs, in such a way that the same requirements are valid for
both. In this case we should call the provider only once.

Add some comments and debug prints, and break the functionality into
a couple of functions.

Implement _process_requirement as a coroutine that yields exactly once
the put_index. When needed, the coroutine receives the callspec that
required it, so it can link it the graph.

Also change the way the default namespace is handled: Now it is not
passed around to the recursive calls of process_requirement. Instead,
it is appended to the correct place in the end, when we know what the
correct namespac for this node is. We construct the default dictionary
meanwhile, and allow to pass it to push_nslevel.

Tests are introduced and modified accordingly.

A couple of missing things are:

 - Remove the put_index stuff from the set_result part. This should
   always have been handled here.

 - A cache for the generated nodes, to not to have to recursively
   build their dependencies each time.

 - A way to specify the ExecModes. In particular we should have a Do
   Nothing mode by default, for the internal nodes, so that we can
   free resources when all the dependencies have used that parameter.
   This can be easily implemented by writing the value in each of the
   output's default namespaces, and then deleting those in set_result.