New format for action parsing
This is breaking change. It is required because with the current
procedure it is not possible to have separate namespace specification
in different action specs.

We now require that everything is a list to begin with. This way we
can keep most of the current parsing logic, and the syntax only
requires a little extra '-'. Incidentally this resembles closely the
actiongroups of smpdf.

Maybe it would be good to have a direct access to previous parser for
simple cases, like just a flat list of actions, but how frequent are
they anyway outside toy examples?