Add remove outer check
It is still to be decided if this is a good idea and how many extra
checks it needs.

Note that at the moment the checks executes, if the parameters are
also arguments of the function, they are required to be in the
namespace, so no exception catching is necessary (it would be
a programming error rather than an user error).

The motivation is that when we want to have different behaviour
according to which elements are not none, we frequently want the inner
ones to "pull" tighter and force ignoring the outer ones. Note that
this only alters the private namespace of the current provider.