Add app.py
This is used to substitute validphys.app. This interface has all the
possible problems of the OOP design, I consider it useful as thinks
like logging or parallel execution and such are much better handled at the reportengine
level, so that client applications do not have to reimplement the
parser each time.

However it is difficult to say exactly how this monster should look
like from only one example (validphys), therefore it will probably
have to be corrected heavily. In particular, now the hooks are placed
in semirandom locations.

In principle, this should have the same behaviour as the old
validphys.app when instanced correctly.