Add colored output
Add a function to highlight the tracebacks (of the critical messages)
and a log handler that can print colors (Unfortunately the interface
sucks and we have to create a new formatter for each message). This
doesn't come for free as it adds two dependencies, pygments and
blessings.

Everyone likes colors.