Implement help system and extra_modules
We customize the argument parser to have a dynamic description, and
the help action to accept an argument about which to ask for help.
It is not so easy to have "validphys help xxx", but instead we have
validphys --help xxx, which is easier to implement and more
consistent.

The help system is implemented using the corresponding helputils
module, which contains utils for correctly formatting and printing
text, parsing the docstrings in the config and adding fancy colors.

In order to categorize better and accept extra_modules, we accept
provider modules as strings and import them using importlib. This
should also allow to delay the imports and offer better startup times
(validphys --help is rather slow).

In order to implement the help for config, we must remember the order
of definition of the parameters, which we do my adding yet another
metaclass.

The help for the providers is still missing.