Improve help system
Enable help for modules and individual actions. This warrants the
introduction of several new APIs.

Explain providers
-----------------

Given a provider name, recursively walk trough all the dependencies:
the other providers, keys known from the config class and unknown
keys. The output is a rather complicated list, but as it turns out
it's easy enough to parse and follow, as demostrated in helputils.

Highlights
----------

The functions that write something to the output folder, now have
a "highlight" property which is used when showing the help (with
different colors and everything). This helps users identify the
providers that are interesting and distinguish them from those that
give intermediate results and are more of an implementation detail.

Wrapping
--------

The textwrap module is broken in many ways, so rather than trying to
force it into something sane, we implement the relevant functionality,
in the sane_wrap function. This respects line breaks unlike textwrap.
Of course it is implemented with a recursive generator.

With all that, the help system prints:
 - For modules: name, docstring and functions that look like providers
   (including docstring and highlight).

 - For actions: module, highlight, docstring, dependencies and
   provider upon which they depend (to one level of recursion).

Ideally, we should have tests for all this...