Rewrite execute_parallel using curio
Now the code is much simpler and more logical, without weird executors
and event loops being passed around.

Also asincio swallows pickling errors and hangs around, which is a no
go.

Now in each step we simply get all runnable nodes, send them to
a pool, and once the results arrive, we iterate.

There are good reasons for not doing this, such as curio being an
experimental library, and introducing an extra dependency (and
recipe), but the above advantages triumph.