Overhaul the way namespaces work
It was causing great trouble. The notion that each part has to take
care of its namespace structure is nice in that one can partially
resolve a namespace but it makes the logic very difficult, and in fact
it was wrong for nested dicts. Now we have a global namespace
structure indexed by tuples. This should be faster in common cases
(quering for an existing namespec) and it's easier to understand.

Now, because there is no self similarity (it only makes sense to call
resolve on the root namespace and it creates difficult to debug
errors to do in other way), an object oriented interface with
a Namespace object makes more sense.