Introduce collect functionality
collect(function, fuzzyspec) will expand fuzzyspcec from the point
where collect is required and generate one normal  node for each resulting
namespace specification, and aggregate the result in a list.

To do that:

 - We introduce node_flags, which tell each node to do something upon
   completion. Once the interface is clear, set_result could be
   written in terms of node flags.

 - The collect node causes each of the nodes it spans to add a flag
   specifying their result to be set to a particular key in
   a odict.

 - The result of the collect node is simply the values of the odict.

 - The same flag can be used to implement filters.

 - We have a new collect node type, which looks the same as the
   callsspec node for the moment. Possibly these nodes will have to
   be full blown objects and grow methods on their own, instead of
   hacking resbuilder and resexecutor in varios disconected places.

 - We split the functionality to push a defaults namespace, which is
   useful for both types of nodes.

 - We have _process_requirement yield the new node as well as the
   put_index.

Help doesn't work for collect nodes at this point, and the mess with
the put_indexes needs a few tests to be clear (hopefully can be
phased out with the node flags, at least partially).