Add argcheck
This makes easier to write what is by far the simplest and most
common way to write checks. The input of the check function is a set
of arguments, and the output is either none or a dictionary with the
updated values of the arguments, after the transformations from the
check.

This is implemented using a saturate function, which calls a function
from the arguments contained in a dictionary.