Allow config to return explicit nodes
This will be useful for reports and such.

There is  slight design problem in that config keys do (and should)
require actions to trigger them, so it's slightly unclear what the
interplay with this feature will be (i.e. report can be implemented
as the identity function taking a template which is a node, but
that's suboptimal).