Enable tables in report
I'm not sure that wrapping them is a great idea, as contrary to the
figures, we possibly want to use tables both as inputs and outputs.
Let's see how well this works in practice.

Otherwise yet another protocol could be added to specify which is the
"return value" and which is the "saved value", but maybe it has to
wait until some reorganization of the namespaces.

Also add pandas as an explicit dependency (doesn't matter too much in
practice).