Don't return Tables by default
The symmetry with Figure is only apparent: we may well want to produce
DataFrames, use them as input to another provider and at the same time
save them. That would require changing the interface in the
user-defined functions that take dataframes as input so they use
table.dataframe (or having a very complicated transparent wrapper over
DataFrame).