Avoid loading Qt in matplotlib
This interacts badly with LHAPDF for some reason, and is completely
unnecessary. If one loads something with LHAPDF after the plotting is
initialized, that results in an error like:

PDF grid data error on line 7: 1 flavor entries seen but 11 expected

which would be good to debug. This also seems to cause LHAPDF to
return incorrect results deterministically (i.e. always the same wrong
result) if the pfds are queried after qt is loaded, so it's a very
serious problem.