Change how Table works again
We need both a fully functioning dataframe and something that knows
how to print itself, including with some metadata like the
description.

A bare class with a dataframe attribute is bad because it breaks the
interfaces of things expecting dataframes.

A bare dataframe cannot contain metadata.

Therefore we return a subclass of Dataframe (that tries to not copy
the data when possible).