Disambiguate targets and fuzzytargets
We were using the same structure to refer to two different things,
which was becoming increasingly confusing. Now there is targets and
fuzzytargets, which have the same relation as nsspecs and fuzzyspecs:
fuzzytargets are expanded into actual targets that get executed.

This will allow to fix some issues with the template parsing and have
a smarter with block that behaves more like a for loop.