Dump environment properties to namespace
Some things like paths are useful for actions that do little but
writing stuff to disk.

While for figures and tables one can make the argument that we are
better off managing things with the complicated mechanism of
preparing paths and so on, so the functions can be used outside
reportengine, this doesn't make sense for everything. For example it
is absurd to have the report function doing absolutely nothing.