Always resolve parameters completely
It is not enough for a variable to already exist in the namespace,
but it must also be in the correct layer of the stack. Otherwise we
have to parse again based on the new parameters.

Now, if we find a result in the namespace we save it and only return
it if we find that it's up to date.

All in all resolve_key and process requirement are two very
complicated functions that are annoyingly similar. Eventually they
should be merged into one somehow.