Replace {@with@} with for loops
Change completely how report works. In several respects this way is
simpler because:

 - The (custom) syntax becomes context free. We don't need any
   external state in order to replace the template with jinja2 tags.
   Also, thanks to how jinja scopes work, we can use only one
   variable ("spec") to have jinja keep track of the state (because
   the value is dropped at the end of the for loop).

 - The report nodes do not need any complicated add_to_dict tags, but work
   exactly the same as normal function nodes.

- We use the already existing namespaces functionality (by exposing
  two small functions) to resolve the target values in the report.
  This is much better than attempting to duplicate this functionality.

The downside is that the node processing becomes too specific to the
report. Ideally we would have an interface that allows creating
custom nodes without having to deal with the complicated code inside
resourcebuilder.

Now with blocks are not simple syntax sugar for repeating target, but
repeat everything inside the block for each nsspec in the fuzzyspec.
This makes writing reports much nicer and simpler.