Fix namespace resolution in report
This fix is not nice in that it makes the interface more complicated
and less logical.

On the other hand all this is scheduled to be overhauled (with proper
Object Oriented nodes as opposed to random bits of functionality
everywhere in the code) so it should be good enough for now.