Fix template nodes inside namespaces
Templates were not written to the namespace at the correct stage.
A template inside a mapping would be processed, but its output dropped
immediately.

The situation is tricky because templates do depend on the nsspec
that comes after the actual template filename is located. Maybe the
better solution would be to make an intermediate provider. The
"template" should instead just deal with the filename.

Also we need to drop the assert that checks that a key was not
already set, which was useful to uncover bugs.

Need to think more on this.