Move report style to its own provider
Main motivation is copying the style once, instead of once per
report. There is a parameter, but it doesn't do much at the moment,
and it's not so clear what should it anyway. Possibilities are:

 - Accept an URL and not copy.
 - See if its an URL, or a file and copy always.
 - Another parameter to specify.
 - ...

I'd prefer CSS to work without an internet connection, but when we
upload the reports we possibly want to link to the same style...