Add option to limit allowed keys
If a Config class defines an allowed keys attribute, processing keys that are
neither parse_<key> functions not in the allowed_keys map will result
in a ConfigError. The values of the allowed_keys map are types which
will be compared to the input value via an isinstance check. If the
Check is False, a BadInputType error will be raised.

Adding something to allowed_keys is different from defining a parse_
function for it, because the value may be parsed recursively (e.g. if
it is a mapping).