Fix bug in from_
This is a difficult one: The resolve_key that converts the input
parameter into a parsed parameter will place it in the outermost
namespace allowed. Since we are not limiting way in any way in some
situations (like when no parser is defined at all) it goes all the
way to the top namespace. We restrict it to not be beyond the key
from which it came in the first place.

The long term fix is to get rid of the put_index craziness, and have
a set() of all the input prams, which always make sense to hash since
an input does not mutate!