Save input card in output
This is very useful now that we upload the reports online. It is
still lacking in that at least the templates should be uploaded as
well. However this requires getting rid of the jinja loader in favour
of our own, that makes available that information (and doesn't read
the file before it's needed).