Add file finder API
This is mostly to replace the awful (for our purposes at least) API of
the jinja2 loaders. It will now become easier to copy the templates to
the input folder in such a way that the input config is fully
reproducible by reruning it.

Also, it is no longer needed to pass around open files sooner than
they are needed.

The new functionality is mainly to be used internally and has 100%
test coverage.