Use the file finder instead of the jinja loader
This preserves information about the file path and allows to copy it
to the input folder (and also we don't have to pass around the
opened template contents).

The preserving functionality is still green (and it should be managed
more centrally and granularly) but should be good enough to be able
to do validphys output/input/runcard.yaml and get the same result.