Fix yet another bug in from_
As it turns out, commit ee2ab3 (that "fixed" another) bug was a pure
piece of garbage. The fix to all the put_index madness is much
simpler: Just parse everything needed from scratch from the input.
This maps exactly the expected from_ semantics and it's much easier
to understand what's going on, instead of trying to decipher the
madness with put_indexes. At least now it starts becoming clear how
the input parsing machinery should work in the first place...