Support bibliography in report
Contrary to templates and meta we require the user application to
specify how to deal with bibliographies. It probably doesn't make
sense to copy them to the report, since they can be unbounded in
size.

At this point it would possibly make sense to have a more pluggable
API for the report. Ideas...?