Allow collect to be resolved at compile time
If all the collected values are known at compile time, we also
resolve the underlying collect.

There are a few reasons why this is a bad idea, like adding technical
debt and making things too implicit, but these will be resolved in
the future. For now this is useful for the validphys dataspecs API.