Add namespaces_ trap
This generates a namespace in the same way as if it was written in
a {@with@ block in the report. This allows to easily construct input
namespaces without having to worry much about the dependencies, which
in turn allows applications to expect that a given namespace has
a given structure, such as fir the dataspecs API in validphys.

This allows to sidewalk the current limitation that nested namespaces
are not lazy. Once they become lazy, this will probably be
a convenient shortcut. On the plus side, it was surprisingly easy to
implement. On the minus, it is something that might complicate
possible future code refactorings. In any case it is needed now.