Add new style action processing
The old action specification was essentially impossible to explain and
gives me a headache. We will maintain it for the foreseeable future
though. However, for now we introduce a better way to write actions,
with the report engine suggestions. It will be automatically suggested
to users of the old action specifications, by showing them the exact
re emplacement.