Take into account all the Config bases for help
Remove the OrderedClass metaclass, as it implements behaviour that is
default since Python 3.6, which is the minimum version that we support.

It wouldn't work unmodified with bases classes, as we have in vp2 now.
Thus the help would only refer to the most derived class, and thus be
meaningless.

Instead, implement a mechanism that lists class attributes in definition
order (cannot be based on inspect.getmembers, which instead gives
alphabetical order).

This change allows to list all the parsing functions and production
rules that are used in a given Config class.